backend/requirements.txt
black==23.7.0
flake8==6.0.0
Flask==3.0.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
gunicorn==22.0.0
isort==5.10.1
marshmallow==3.21.1
mypy==1.4.1
psycopg2-binary==2.9.9
pytest==8.1.1
sqlalchemy-stubs==0.4
Werkzeug==3.0.3
